```
benchmark                       old ns/op     new ns/op     delta
BenchmarkApplyNewObject-12      1355981       1363050       +0.52%
BenchmarkUpdateNewObject-12     1565341       1228947       -21.49%
BenchmarkRepeatedUpdate-12      443528        396604        -10.58%
BenchmarkSetToFields-12         5154          5107          -0.91%
BenchmarkFieldsToSet-12         10050         10170         +1.19%
BenchmarkYAMLToTyped-12         56860         54439         -4.26%

benchmark                       old allocs     new allocs     delta
BenchmarkApplyNewObject-12      4264           4214           -1.17%
BenchmarkUpdateNewObject-12     4850           3661           -24.52%
BenchmarkRepeatedUpdate-12      1255           1179           -6.06%
BenchmarkSetToFields-12         14             14             +0.00%
BenchmarkFieldsToSet-12         82             82             +0.00%
BenchmarkYAMLToTyped-12         90             90             +0.00%
```