@lavalamp 2019-08-17
Didn't just skip the tests because if the code is checked in, it needs to not bitrot. Otherwise I need to delete it.